updating to sequence kit 4.3.0
