+++ package.json
@@ -10,18 +10,18 @@
     "postinstall": "(cp -n .env.example .env) || echo already exists"
   },
   "dependencies": {
-    "0xsequence": "2.0.6",
-    "@0xsequence/design-system": "^1.7.6",
-    "@0xsequence/indexer": "^2.0.6",
-    "@0xsequence/kit": "4.0.3",
-    "@0xsequence/network": "^2.0.6",
-    "@tanstack/react-query": "^5.51.23",
-    "ethers": "6.13.2",
-    "next": "14.2.5",
+    "0xsequence": "2.0.11",
+    "@0xsequence/design-system": "^1.7.8",
+    "@0xsequence/indexer": "^2.0.11",
+    "@0xsequence/kit": "^4.3.0",
+    "@0xsequence/network": "^2.0.11",
+    "@tanstack/react-query": "^5.59.0",
+    "ethers": "6.13.3",
+    "next": "14.2.14",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "viem": "^2.19.4",
-    "wagmi": "^2.12.5"
+    "viem": "^2.21.18",
+    "wagmi": "^2.12.16"
   },
   "devDependencies": {
     "@types/node": "^22.3.0",
@@ -30,12 +30,12 @@
     "@typescript-eslint/eslint-plugin": "^7.18.0",
     "@typescript-eslint/parser": "^7.18.0",
     "eslint": "^8.57.0",
-    "eslint-config-next": "14.2.5",
+    "eslint-config-next": "14.2.14",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-unused-imports": "^4.1.3",
     "prettier": "^3.3.3",
-    "typescript": "^5.5.4"
+    "typescript": "^5.6.2"
   },
   "packageManager": "pnpm@9.5.0+sha512.140036830124618d624a2187b50d04289d5a087f326c9edfc0ccd733d76c4f52c3a313d4fc148794a2a9d81553016004e6742e8cf850670268a7387fc220c903"
 }
