+++ src/app/layout.tsx
@@ -1,17 +1,22 @@
-import type { Metadata } from "next";
+"use client"
+import { SequenceKit } from "@0xsequence/kit";
 import "./globals.css";
-import Providers from "./Providers";
-
-export const metadata: Metadata = {
-  title: "Embedded Wallet Nextjs Boilerplate Demo",
-  description: "Embedded Wallet Nextjs Boilerplate Demo",
-};
+import "@0xsequence/design-system/styles.css";
+import { useEffect, useState } from "react";
+import FullScreenLoading from "./components/FullScreenLoading";
+import { config } from "../config";
 
 export default function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+  const [isClient, setIsClient] = useState(false);
+
+  useEffect(() => {
+    setIsClient(true);
+  }, []);
+
   return (
     <html lang="en">
       <head>
@@ -53,9 +58,13 @@ export default function RootLayout({
             ></path>
           </svg>
         </a>
-        <Providers>
-          <div id="root">{children}</div>
-        </Providers>
+        {!isClient ? (
+          <FullScreenLoading />
+        ) : (
+          <SequenceKit config={config}>
+            <div id="root">{children}</div>
+          </SequenceKit>
+        )}
       </body>
     </html>
   );
