+++ src/chains.ts
@@ -1,0 +1,12 @@
+import {
+  arbitrumSepolia,
+  Chain,
+  mainnet,
+  polygon,
+  polygonAmoy,
+} from "wagmi/chains";
+
+export const chains = [mainnet, polygon, polygonAmoy, arbitrumSepolia] as [
+  Chain,
+  ...Chain[],
+];
