+++ src/metadata.ts
@@ -1,0 +1,6 @@
+import type { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "Embedded Wallet Nextjs Boilerplate Demo",
+  description: "Embedded Wallet Nextjs Boilerplate Demo",
+};
